new event page, routing changes, Menu layout
