+++ components/frontend-web/src/app/app-routing.module.ts
@@ -6,12 +6,16 @@ const routes: Routes = [
 
   { path: '', redirectTo: 'ui/landing', pathMatch: 'full' },
   { path: 'ui/landing', loadChildren: './pages/landing/landing.module#LandingPageModule' },
-  { path: 'ui/event', loadChildren: './pages/event/event.module#EventPageModule', runGuardsAndResolvers: 'always', canActivate: [AuthGuard]},
   { path: 'ui/login', loadChildren: './pages/login/login.module#LoginPageModule' },
+  // { path: 'ui/event', loadChildren: './pages/event/event.module#EventPageModule'},
+  { path: 'ui/event/:eventId', loadChildren: './pages/event/event.module#EventPageModule'},
+  { path: 'ui/create-event', loadChildren: './pages/create-event/create-event.module#CreateEventPageModule' },
   { path: 'ui/playlist-user', loadChildren: './pages/playlist/playlist.module#PlaylistPageModule', runGuardsAndResolvers: 'always' , canActivate: [AuthGuard]},
   { path: 'ui/playlist-curator', loadChildren: './pages/playlist/playlist.module#PlaylistPageModule', runGuardsAndResolvers: 'always' , canActivate: [AuthGuard]},
   { path: 'ui/playlist-event', loadChildren: './pages/playlist/playlist.module#PlaylistPageModule', runGuardsAndResolvers: 'always' , canActivate: [AuthGuard]},
-  { path: ':userEventID', loadChildren: './pages/playlist/playlist.module#PlaylistPageModule', runGuardsAndResolvers: 'always' , canActivate: [AuthGuard] }
+  // { path: ':userEventID', loadChildren: './pages/playlist/playlist.module#PlaylistPageModule', runGuardsAndResolvers: 'always' , canActivate: [AuthGuard] },
+  { path: ':eventID', redirectTo: 'ui/event/:eventID', pathMatch: 'full' },
+  { path: '**', redirectTo: 'ui/landing', pathMatch: 'full' }
 ];
 
 @NgModule({
