+++ components/frontend-web/src/app/app.component.html
@@ -1,7 +1,7 @@
 <ion-app>
 
     <ion-menu>
-        <ion-content fullscreen>
+        <ion-content fullscreen color="dark">
 
             <ion-header>
                 <ion-toolbar color="dark">
@@ -9,48 +9,61 @@
                 </ion-toolbar>
             </ion-header>
 
-            <ion-list no-padding>
+            <ion-list no-padding color="dark">
+
+                <ion-menu-toggle autoHide="false" *ngIf="!userState?.isLoggedIn">
+                    <ion-item [routerDirection]="'root'" routerLink="/ui/landing" color="opendj">
+                        <ion-icon slot="start" [name]="'home'"></ion-icon>
+                        <ion-label>Start</ion-label>
+                    </ion-item>
+                </ion-menu-toggle>
+
                 <ion-menu-toggle *ngIf="userState?.isLoggedIn" autoHide="false">
-                    <ion-item color="medium">
+                    <ion-item color="dark">
                         <ion-icon slot="start" name="contact"></ion-icon>
                         <ion-label>
                             {{userState.username}}
                         </ion-label>
-                        <ion-icon slot="end" name="star" *ngIf="userState?.isCurator && !userState?.isEventOwner" color="yellow"></ion-icon>
+                        <ion-icon slot="end" name="star" *ngIf="userState?.isCurator && !userState?.isEventOwner"
+                            color="yellow"></ion-icon>
                         <ion-icon slot="end" name="star" *ngIf="userState?.isEventOwner" color="red"></ion-icon>
                     </ion-item>
                 </ion-menu-toggle>
 
                 <ion-menu-toggle auto-hide="false">
-                    <ion-item *ngIf="userState?.isLoggedIn" [routerDirection]="'root'" routerLink="/ui/playlist-user" color="light">
+                    <ion-item *ngIf="userState?.isLoggedIn" [routerDirection]="'root'" routerLink="/ui/playlist-user"
+                        color="opendj">
                         <ion-icon slot="start" [name]="'musical-notes'"></ion-icon>
                         <ion-label>User View</ion-label>
                     </ion-item>
-                    <ion-item *ngIf="userState?.isCurator" [routerDirection]="'root'" routerLink="/ui/playlist-curator" color="light">
+                    <ion-item *ngIf="userState?.isCurator" [routerDirection]="'root'" routerLink="/ui/playlist-curator"
+                        color="opendj">
                         <ion-icon slot="start" [name]="'navigate'"></ion-icon>
                         <ion-label>Curator View</ion-label>
                     </ion-item>
 
-                    <ion-item *ngIf="userState?.isEventOwner" [routerDirection]="'root'" routerLink="/ui/event" color="light">
+                    <ion-item *ngIf="userState?.isEventOwner" [routerDirection]="'root'" routerLink="/ui/event"
+                        color="opendj">
                         <ion-icon slot="start" [name]="'options'"></ion-icon>
                         <ion-label>Edit Event</ion-label>
                     </ion-item>
 
                 </ion-menu-toggle>
 
-                <ion-menu-toggle autoHide="false">
-                    <ion-item button (click)="logout()" color="light">
+                <ion-menu-toggle autoHide="false" *ngIf="userState?.isLoggedIn">
+                    <ion-item button (click)="logout()" color="opendj">
                         <ion-icon slot="start" name="log-out"></ion-icon>
                         <ion-label>
                             Logout
                         </ion-label>
                     </ion-item>
                 </ion-menu-toggle>
+
             </ion-list>
         </ion-content>
     </ion-menu>
     <ion-router-outlet main [swipeGesture]="false"></ion-router-outlet>
-<!--
+    <!--
     <div *ngIf="platform.is('desktop')" class="desktop-footer" padding>
         <a href="https://www.openshift.com/" title="Powered by OpenShift Online">
             <img alt="Powered by OpenShift Online" title="Powered by OpenShift Online" src="https://www.openshift.com/images/logos/powered_by_openshift_reverse.png" />
