+++ components/frontend-web/src/app/app.component.ts
@@ -14,6 +14,7 @@ import { UserSessionState } from './models/usersessionstate';
 })
 export class AppComponent implements OnInit {
 
+  // userState is important for displaying the menu options
   userState;
 
   constructor(
@@ -26,11 +27,12 @@ export class AppComponent implements OnInit {
     private userDataService: UserDataService
   ) {
     this.initializeApp();
+    this.registerEventSubscribers();
   }
 
   initializeApp() {
     this.platform.ready().then((readySource) => {
-      console.log(`Platform: ${readySource}`);
+      console.debug(`Running on Platform: ${readySource}`);
 
       if (readySource === 'cordova') {
         this.statusBar.styleDefault();
@@ -44,9 +46,13 @@ export class AppComponent implements OnInit {
     this.userState = await this.userDataService.getUser();
   }
 
-  private async saveUserState() {
-    console.debug('saveUserState');
-    await this.userDataService.updateUser(this.userState);
+  registerEventSubscribers() {
+    console.debug('registerEventSubscribers');
+
+    this.events.subscribe('sessionState:modified', state => {
+      console.debug('Received sessionState:modified event');
+      this.userState = state;
+    });
   }
 
   logout() {
@@ -57,11 +63,6 @@ export class AppComponent implements OnInit {
   async ngOnInit() {
     console.debug('begin ngOnInit()');
     await this.loadUserState();
-
-    this.events.subscribe('user:modified', newUser => {
-      console.debug('Received user:modified event');
-      this.userState = newUser;
-    });
   }
 
 }
