+++ components/frontend-web/src/app/pages/create-event/create-event.module.ts
@@ -1,0 +1,28 @@
+import { NgModule } from '@angular/core';
+import { SharedModule } from 'src/app/components/shared.module';
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+import { Routes, RouterModule } from '@angular/router';
+
+import { IonicModule } from '@ionic/angular';
+
+import { CreateEventPage } from './create-event.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: CreateEventPage
+  }
+];
+
+@NgModule({
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    SharedModule,
+    IonicModule,
+    RouterModule.forChild(routes)
+  ],
+  declarations: [CreateEventPage]
+})
+export class CreateEventPageModule {}
