+++ components/frontend-web/src/app/pages/create-event/create-event.page.scss
@@ -20,6 +20,4 @@
     color: #fff;
 }
 
-
-ion-badge { margin-right: 10px;}
-ion-card-title { color: #fff;}
+ion-item{ --border-color: #333; }
