--- components/frontend-web/src/app/providers/env.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { EnvService } from './env.service';
-
-describe('EnvService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: EnvService = TestBed.get(EnvService);
-    expect(service).toBeTruthy();
-  });
-});
